Contracts refactor phase 1: extract shared models + ContractDisplayService
Moves the Contracts page's shared type block and pure display/format
helpers (truncate, formatSompiToKas, explorer links, status/source
labels, buildShareLink) out of ContractsPageComponent into
contracts-page.models.ts and ContractDisplayService. The component
keeps identical-signature delegating wrappers, so the 3.8k-line
template needed zero edits. First step of splitting the 8.5k-line
ContractsPageComponent into focused pieces (see plan).
